refactor: rename-smart-contracts-dir-to-contracts
