+++ .gitignore
@@ -32,13 +32,12 @@ lcov.info
 # Hardhat Ignition default folder for deployments against a local node
 ignition/deployments/chain-31337
 
-smart-contracts/lib/
-
-smart-contracts/artifacts/
-smart-contracts/cache/
-smart-contracts/coverage.json
-smart-contracts/typechain-types/
-smart-contracts/broadcast/**/run-*.json
+contracts/lib/
+contracts/artifacts/
+contracts/cache/
+contracts/coverage.json
+contracts/typechain-types/
+contracts/broadcast/**/run-*.json
 
 *.csv
 
@@ -49,4 +48,3 @@ smart-contracts/broadcast/**/run-*.json
 .env.**
 
 .vercel
-
