+++ .gitmodules
@@ -1,9 +1,9 @@
-[submodule "smart-contracts/lib/forge-std"]
-	path = smart-contracts/lib/forge-std
+[submodule "contracts/lib/forge-std"]
+	path = contracts/lib/forge-std
 	url = https://github.com/foundry-rs/forge-std
-[submodule "smart-contracts/lib/openzeppelin-contracts"]
-	path = smart-contracts/lib/openzeppelin-contracts
+[submodule "contracts/lib/openzeppelin-contracts"]
+	path = contracts/lib/openzeppelin-contracts
 	url = https://github.com/OpenZeppelin/openzeppelin-contracts
-[submodule "smart-contracts/lib/openzeppelin-contracts-upgradeable"]
-	path = smart-contracts/lib/openzeppelin-contracts-upgradeable
+[submodule "contracts/lib/openzeppelin-contracts-upgradeable"]
+	path = contracts/lib/openzeppelin-contracts-upgradeable
 	url = https://github.com/OpenZeppelin/openzeppelin-contracts-upgradeable
