+++ contracts/src/Vote.sol
