+++ contracts/src/interfaces/ICitizen.sol
